Pulled objects out of DBData, created a seperate file for each, and deleted DBData.
