+++ server/conf/application.conf
@@ -285,7 +285,7 @@ play.filters {
   # This is useful to prevent cache poisoning attacks.
   hosts {
     # Allow requests to example.com, its subdomains, and localhost:9000.
-    allowed = ["pandora06.cs.trinity.edu:9005", "pandora06.cs.trinity.edu:9006", "pandora06.cs.trinity.edu:9007", "pandora06.cs.trinity.edu:9008", "pandora06.cs.trinity.edu:9009", "localhost:9005"]
+    allowed = ["pandora06.cs.trinity.edu:9005", "pandora06.cs.trinity.edu:9006", "pandora06.cs.trinity.edu:9007", "pandora06.cs.trinity.edu:9008", "pandora06.cs.trinity.edu:9009", "localhost:9005", "localhost:9000"]
   }
 }
